+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.0.4] - 2023-3-12
+
+### Added
+ - Icon in the project browser when a file has project browser
+
+### Changed
+ - Only files that have been moved, imported or deleted will be checked using OnPostprocessAllAssets
+ - Files will now be checked in a background task
+ - AssetChangesdetector has been renamed to FaultyAssetHandler
+
+### Fixed
+ - Deleted GUIDs will get removed from rules
+
 [1.0.3] - 2023-2-20
 
 ### Added